requirements.txt
clipspy==1.0.6
customtkinter==5.2.2
